Simplify API w.r.t. string types and parallelism
This drops the parallel Rust entry points as they offer little over direct usage
of Rayon while forcing a copy and requiring the corpus to be available as a
slice. The corresponding Python entry points are kept for convenience, but are
implemented directly using Rayon instead.

Further, all string types are standardized on `&str` thereby avoiding
unnecessary allocations when the data is already available in string form.
Furthermore, the Python entry points can directly borrow from Python-managed
string data via PyO3's `PyBackedStr`.

This is obviously not backwards-compatible on the Rust side but I think it is a
simpler yet more efficient API to present there and hence wanted to at least
propose this even though the difference is probably not worth a major version
bump.