WIP: Use Finite State Transducers (FST) as the backing store for language models
FST have the nice properties of both compressing the ngram data by exploiting
common suffixes and prefixes as well as being embeddable into the binary in a
form that is directly suitable for look-up thereby avoiding the separate
decompression step and indireclty using memory mappings as supplied by the
operating system for all binaries.

This is still WIP since I do not know how to regenerate the language models and
it also seems like the unique ngram models are built elsewhere.

TODO:
* Integrate unique ngram models
* Regenerate all language models
* Drop non-unified models